Add portfolio categorized by year
This commit categorizes images by year using their
name to determine year using a simple regex:
- /^(\d{4}).*/

Thus, if an image filename doesn't match this regex,
it will not be added to the portfolio.

A new Vue component is now implemented: `Portfolio`.
This one has been created in `Portfolio` folder
with `ImageList` & `ImageItem` for more consistency.