Add expanded image behavior on click over images
This commit adds the feature about expand images
when user clicks over an image of portfolio.

This feature uses a new Vuex Store in order to
notify `ExpandedImage.vue` component that an
image has to be expanded. Then, an other click
event is bound on the expanded image in order
to hide expanded image. It's a toggle behavior.