Update image grid behavior
This commit updates portfolio images grid.

To respect basic dimensions of portfolio's images,
we opted for a JS computed solution to render images
with their right dimension instead of background image
CSS property as "cover" value.

It could be implemented with CSS but it was not flexible
enough and some visual bugs wasn't covered.

Thus, we chose "masonry" VueJS plugin.
@see https://github.com/shershen08/vue-masonry based on
https://masonry.desandro.com/ lib.