Add new Victor Vaysse's CV page
This commit adds Victor Vaysse's CV to a new
page.

Note: this page is already routed with previous
the previous commit.

It uses a new component `Category`. This component
consumes props to display the CV categories.
Categories are main parts of a CV, for example:
- experiences,
- skills,
- etc.

It is easy to add or remove any words of this CV
just by updating `categories` state of `pages/cv.vue`
page.