Resize portfolio images in a smaller size
This commit resizes all images in portfolio
in a smaller sizes. Indeed, big images are too
expansives for low users connections.

A script in API is now enabled to resize a folder
or a file using a command-line:
- npm run resizer -- [options]
@see README.md for more information

Note: previous npm commands have been updated in
order to allow build in production environment.