Fix design behavior on old browsers
The main feature of this commit is to
resolve some issues about design about
old browser versions. Especially for
safari on iOS.

It also add the Victor Vaysse mail in top
of 2 pages for a better visibility.

Finally, it adds a link on the logo title
for the home page.