Add lazyload behavior on images
This commit refactors the way masonry is used.
Henceforth, we no longer use the JS vue plugin
because we want to manage more precisely how
the lazyload should work.

Thus, we use the "masonry-layout" library directly
to instantiate the framework. In addition we have
added a new library "view-in-viewport-mixin" which
allows to define if a component is in the viewport
or not.

Note: lazyload is not enabled for the first item
list.